requirements.txt
gitdb==4.0.11
GitPython==3.1.43
requests==2.31.0
selenium==4.20.0
urllib3==2.2.1
